12/host computer/maze_explorer/requirements.txt
# BLE Communication
bleak>=0.19.0

# Numerical Computing
numpy>=1.20.0

# SLAM Algorithm (used in Stage 2)
# Note: BreezySLAM is not on PyPI, install from GitHub:
# pip install git+https://github.com/simondlevy/BreezySLAM.git#subdirectory=python

# Visualization (used in Stage 2)
matplotlib>=3.3.0

# Image Processing (map saving)
Pillow>=8.0.0

